+++ src/i18n.ts
@@ -15,7 +15,7 @@ export const localeLanguage: TLanguage = ((): TLanguage => {
     if (language === 'en') return 'en';
     if (language === 'en-US') return 'en';
     if (language === 'zh-cn') return 'zh-cn';
-    if (language === 'zh-tw') return 'zh-cn'; // TODO: if all languages is call zh-cn, then make it
+    if (language === 'zh-tw') return 'zh-cn'; // if all languages is call zh-cn, then make it TODO
 
     return 'en';
 })();